Realised that if the server went down, the user wouldnt know.
So I added an API in the server to give the status of the server.

I also added in a singleton service in the App to constantly ping the server every 10s to check its health. if it goes down, it logs out the user and informs them to try again.